+++ src/main/java/de/codescape/jira/plugins/scrumpoker/workflow/StartSessionWorkflowFunction.java
@@ -2,6 +2,8 @@ package de.codescape.jira.plugins.scrumpoker.workflow;
 
 import com.atlassian.jira.issue.MutableIssue;
 import com.atlassian.jira.workflow.function.issue.AbstractJiraFunctionProvider;
+import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
+import com.atlassian.sal.api.message.I18nResolver;
 import com.opensymphony.module.propertyset.PropertySet;
 import de.codescape.jira.plugins.scrumpoker.service.ErrorLogService;
 import de.codescape.jira.plugins.scrumpoker.service.EstimateFieldService;
@@ -24,12 +26,15 @@ public class StartSessionWorkflowFunction extends AbstractJiraFunctionProvider {
     private final ScrumPokerLicenseService scrumPokerLicenseService;
     private final EstimateFieldService estimateFieldService;
     private final ErrorLogService errorLogService;
+    private final I18nResolver i18nResolver;
 
     @Autowired
-    public StartSessionWorkflowFunction(ScrumPokerSessionService scrumPokerSessionService,
+    public StartSessionWorkflowFunction(@ComponentImport I18nResolver i18nResolver,
+                                        ScrumPokerSessionService scrumPokerSessionService,
                                         ScrumPokerLicenseService scrumPokerLicenseService,
                                         EstimateFieldService estimateFieldService,
                                         ErrorLogService errorLogService) {
+        this.i18nResolver = i18nResolver;
         this.scrumPokerSessionService = scrumPokerSessionService;
         this.scrumPokerLicenseService = scrumPokerLicenseService;
         this.estimateFieldService = estimateFieldService;
@@ -40,8 +45,7 @@ public class StartSessionWorkflowFunction extends AbstractJiraFunctionProvider {
     public void execute(Map transientVars, Map args, PropertySet propertySet) {
         // check license
         if (!scrumPokerLicenseService.hasValidLicense()) {
-            errorLogService.logError("Unable to start Scrum Poker session because of license errors: "
-                + scrumPokerLicenseService.getLicenseError());
+            errorLogService.logError(i18nResolver.getText(scrumPokerLicenseService.getLicenseError()));
             return;
         }