+++ docs/changelog.md
@@ -14,6 +14,7 @@ category: Administration
 * refactor: extract license check into separate reusable service class
 * feat: add support for Jira 8.13.1
 * test: allow underscores in message bundle for internationalization
+* feat: improve license check across all actions and functions (#86 and #88)
 
 ### [20.10.1] - 2020-10-29